# server/swagger.yaml
openapi: 3.0.3
info:
  title: SuperAppBet SDUI API
  version: 1.0.0
  description: |
    API para o sistema Server-Driven UI do SuperAppBet - uma aplicação de apostas com arquitetura de micro frontends.
    
    Esta API oferece:
    - 🎮 **Gerenciamento de Módulos**: Ativar/desativar funcionalidades dinamicamente
    - 🎨 **Server-Driven UI**: Configurar interfaces remotamente
    - 📱 **Componentes Dinâmicos**: Criar e gerenciar componentes UI
    - 🖥️ **Telas Configuráveis**: Definir layouts de tela via servidor
    - ⚡ **WebSocket Real-time**: Atualizações em tempo real
    - 📊 **Analytics**: Métricas de uso e performance
    
  contact:
    name: SuperAppBet Team
    email: tech@superappbet.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://api.superappbet.com/api
    description: Production Server
...
paths:
...
  /modules:
    get:
      tags: [Modules]
      summary: Listar todos os módulos disponíveis
      description: Retorna lista completa de módulos registrados no sistema
      responses:
        '200':
          description: Lista de módulos retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Module'
                  count:
                    type: integer
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Modules]
      summary: Criar novo módulo
      description: Registra um novo módulo no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateModuleRequest'
      responses:
        '200':
          description: Módulo criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/enabled:
    get:
      tags: [Modules]
      summary: Listar módulos habilitados
      description: Retorna módulos habilitados para um usuário/segmento específico
      parameters:
        - name: userId
          in: query
          description: ID do usuário
          schema:
            type: string
        - name: segment
          in: query
          description: Segmento do usuário
          schema:
            type: string
            default: default
        - name: region
          in: query
          description: Região do usuário
          schema:
            type: string
            default: BR
      responses:
        '200':
          description: Módulos habilitados retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EnabledModule'
                  metadata:
                    $ref: '#/components/schemas/QueryMetadata'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/enable/{moduleId}:
    post:
      tags: [Modules]
      summary: Habilitar módulo
      description: Habilita um módulo específico para usuário/segmento
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EnableModuleRequest'
      responses:
        '200':
          description: Módulo habilitado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/disable/{moduleId}:
    post:
      tags: [Modules]
      summary: Desabilitar módulo
      description: Desabilita um módulo específico
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DisableModuleRequest'
      responses:
        '200':
          description: Módulo desabilitado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/{moduleId}/config:
    get:
      tags: [Modules]
      summary: Obter configuração do módulo
      description: Retorna a configuração atual de um módulo específico
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
        - name: userId
          in: query
          description: ID do usuário
          schema:
            type: string
        - name: segment
          in: query
          description: Segmento
          schema:
            type: string
            default: default
        - name: region
          in: query
          description: Região
          schema:
            type: string
            default: BR
      responses:
        '200':
          description: Configuração retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags: [Modules]
      summary: Atualizar configuração do módulo
      description: Atualiza a configuração de um módulo específico
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateConfigRequest'
      responses:
        '200':
          description: Configuração atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/{moduleId}/actions/{actionType}:
    post:
      tags: [Modules]
      summary: Executar ação do módulo
      description: Executa uma ação específica de um módulo
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
        - name: actionType
          in: path
          required: true
          description: Tipo da ação
          schema:
            type: string
            enum: [place-bet, add-to-betslip, remove-from-betslip, play-game, deposit, withdraw]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExecuteActionRequest'
      responses:
        '200':
          description: Ação executada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/{moduleId}/analytics:
    get:
      tags: [Modules, Analytics]
      summary: Obter analytics do módulo
      description: Retorna métricas de uso e performance de um módulo
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
        - name: startDate
          in: query
          description: Data inicial (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Data final (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: segment
          in: query
          description: Segmento
          schema:
            type: string
        - name: region
          in: query
          description: Região
          schema:
            type: string
      responses:
        '200':
          description: Analytics retornados com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    $ref: '#/components/schemas/ModuleAnalytics'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /modules/{moduleId}/feature-flag:
    post:
      tags: [Modules]
      summary: Configurar feature flag
      description: Define ou atualiza uma feature flag para um módulo
      parameters:
        - name: moduleId
          in: path
          required: true
          description: ID do módulo
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FeatureFlagRequest'
      responses:
        '200':
          description: Feature flag configurada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'

  # SDUI Configuration
...
  /sdui/components:
    get:
      tags: [SDUI, Components]
      summary: Obter biblioteca de componentes
      description: Retorna todos os componentes disponíveis para SDUI
      responses:
        '200':
          description: Biblioteca de componentes retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SDUIComponent'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [SDUI, Components]
      summary: Registrar novo tipo de componente
      description: Registra um novo tipo de componente no sistema SDUI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterComponentRequest'
      responses:
        '200':
          description: Componente registrado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /components:
    get:
      tags: [Components]
      summary: Listar todos os componentes
      description: Retorna lista de componentes com filtros opcionais
      parameters:
        - name: category
          in: query
          description: Filtrar por categoria
          schema:
            type: string
            enum: [promotional, betting, casino, layout, navigation]
        - name: platform
          in: query
          description: Filtrar por plataforma
          schema:
            type: string
            enum: [ios, android, web, all]
        - name: status
          in: query
          description: Filtrar por status
          schema:
            type: string
            enum: [active, inactive, deleted]
            default: active
      responses:
        '200':
          description: Lista de componentes retornada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Component'
                  count:
                    type: integer
                  filters:
                    type: object
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags: [Components]
      summary: Criar novo componente
      description: Cria um novo componente no sistema
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComponentRequest'
      responses:
        '200':
          description: Componente criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        timestamp:
          type: string
          format: date-time

    # Module Schemas
    Module:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [betting, casino, lottery, payments, core]
        components:
          type: array
          items:
            type: string
        screens:
          type: array
          items:
            type: string
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ModuleAction'
        dependencies:
          type: array
          items:
            type: string
        defaultConfig:
          type: object
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        status:
          type: string
          enum: [active, inactive, deprecated]
    EnabledModule:
      allOf:
        - $ref: '#/components/schemas/Module'
        - type: object
          properties:
            config:
              type: object
            enabledAt:
              type: string
              format: date-time
            enabledBy:
              type: string
    ModuleAction:
      type: object
      properties:
        type:
          type: string
        handler:
          type: string
        description:
          type: string
        parameters:
          type: object
    CreateModuleRequest:
      type: object
      required: [name, version, description, category]
      properties:
        name:
          type: string
        version:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [betting, casino, lottery, payments, core]
        components:
          type: array
          items:
            type: string
          default: []
        screens:
          type: array
          items:
            type: string
          default: []
        actions:
          type: array
          items:
            $ref: '#/components/schemas/ModuleAction'
          default: []
        dependencies:
          type: array
          items:
            type: string
          default: []
        defaultConfig:
          type: object
          default: {}
    EnableModuleRequest:
      type: object
      properties:
        userId:
          type: string
        segment:
          type: string
          default: default
        region:
          type: string
          default: BR
        config:
          type: object
          default: {}
    DisableModuleRequest:
      type: object
      properties:
        userId:
          type: string
        segment:
          type: string
          default: default
        region:
          type: string
          default: BR
        reason:
          type: string
    UpdateConfigRequest:
      type: object
      required: [config]
      properties:
        config:
          type: object
        segment:
          type: string
          default: default
        region:
          type: string
          default: BR
    ExecuteActionRequest:
      type: object
      required: [payload, userId]
      properties:
        payload:
          type: object
        userId:
          type: string
        metadata:
          type: object
          default: {}
    ActionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        metadata:
          type: object
          properties:
            moduleId:
              type: string
            actionType:
              type: string
            executedAt:
              type: string
              format: date-time
    FeatureFlagRequest:
      type: object
      required: [flagName, enabled]
      properties:
        flagName:
          type: string
        enabled:
          type: boolean
        conditions:
          type: object
          default: {}
    ModuleAnalytics:
      type: object
      properties:
        moduleId:
          type: string
        name:
          type: string
        period:
          $ref: '#/components/schemas/DatePeriod'
        metrics:
          $ref: '#/components/schemas/UsageMetrics'
        performance:
          $ref: '#/components/schemas/PerformanceMetrics'

    # SDUI Schemas
...
    SDUIComponent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        schema:
          type: object
        defaultProps:
          type: object
        category:
          type: string
        variants:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
...
    RegisterComponentRequest:
      type: object
      required: [name, schema]
      properties:
        name:
          type: string
        schema:
          type: object
        defaultProps:
          type: object
          default: {}
        category:
          type: string
...
    Component:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [promotional, betting, casino, layout, navigation]
        schema:
          type: object
        defaultProps:
          type: object
        variants:
          type: array
          items:
            type: string
        platforms:
          type: array
          items:
            type: string
        conditions:
          type: object
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
        status:
          type: string
          enum: [active, inactive, deleted]
        version:
          type: integer
    CreateComponentRequest:
      type: object
      required: [name, type, description, category, schema]
      properties:
        name:
          type: string
        type:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [promotional, betting, casino, layout, navigation]
        schema:
          type: object
        defaultProps:
          type: object
          default: {}
        variants:
          type: array
          items:
            type: string
          default: ["default"]
        platforms:
          type: array
          items:
            type: string
          default: ["ios", "android"]
        conditions:
          type: object
          default: {}
...
    DatePeriod:
      type: object
      properties:
        startDate:
          type: string
          format: date-time
        endDate:
          type: string
          format: date-time
    UsageMetrics:
      type: object
      properties:
        activeUsers:
          type: integer
        sessionsCount:
          type: integer
        averageSessionDuration:
          type: integer
        actionCount:
          type: integer
    PerformanceMetrics:
      type: object
      properties:
        averageResponseTime:
          type: integer
        errorRate:
          type: string
        throughput:
          type: integer
        availability:
          type: string
    QueryMetadata:
      type: object
      properties:
        userId:
          type: string
        segment:
          type: string
        region:
          type: string
        appliedAt:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Parâmetros inválidos
            timestamp: "2023-12-01T10:00:00Z"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Recurso não encontrado
            timestamp: "2023-12-01T10:00:00Z"
    ServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: Erro interno do servidor
            timestamp: "2023-12-01T10:00:00Z"
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key para autenticação de endpoints administrativos

    AdminAuth:
      type: apiKey
      in: header
      name: X-Admin-User
      description: Identificação do usuário administrador para auditoria

# WebSocket Events Documentation
...
security:
  - ApiKeyAuth: []
  - AdminAuth: []